renamed rest_api to rpc
