+++ core/src/lib.rs
@@ -15,7 +15,7 @@
 )]
 
 pub mod core;
-pub mod rest_api;
+pub mod rpc;
 pub mod strategies;
 
 #[cfg(test)]
