+++ core/src/rpc/control_panel.rs
